+++ lego/CHANGELOG.md
@@ -1,4 +1,24 @@
 <!-- https://developers.home-assistant.io/docs/add-ons/presentation#keeping-a-changelog -->
+## 0.2.0
+
+### New Features
+
+- Add `--renew-hook` to automatically restart `Home-Assistant core` when renewing certificate.
+
+### Improvements
+
+- Better check for specific domain certificate, issue each domain severalty.
+- Remove config option `interval` in favor of running in specific time. 
+- Change port for the `http challenge` from `80` to high port `8091` to free port `80` for other services.
+- Change default renew time to `30` days to reduce the renew frequency
+
+### Bug Fix
+
+- Fix command selection done once on addon start.
+
+### Others
+
+- Code cleanup
 
 ## 0.1.3